Refactor to get the createExtension functionality in one place
